+++ src/core/useCases/delete-user/interface.go
@@ -1,0 +1,7 @@
+package delete_user
+
+type (
+	DeleteUserUseCase interface {
+		DeleteUser(id string) error
+	}
+)