+++ src/infra/integrations/grpc/client/client.interface.go
@@ -1,0 +1,9 @@
+package grpc_client
+
+import "google.golang.org/grpc"
+
+type (
+	GRPCClient interface {
+		GetConnection(host string) *grpc.ClientConn
+	}
+)