+++ src/infra/integrations/http/jsonplaceholder/main.go
@@ -3,6 +3,7 @@ package jsonplaceholder
 import (
 	ports "go-api/src/core/ports"
 	http_service "go-api/src/main/module/http/client"
+	"os"
 )
 
 type JsonPlaceholderIntegration struct {
@@ -10,10 +11,10 @@ type JsonPlaceholderIntegration struct {
 	rootUrl string
 }
 
-func New(http http_service.HttpClient, root_url string) ports.JsonPlaceholderIntegration {
+func New(http http_service.HttpClient) ports.JsonPlaceholderIntegration {
 
 	return &JsonPlaceholderIntegration{
 		Http:    http,
-		rootUrl: root_url,
+		rootUrl: os.Getenv("JSON_PLACE_OLDER_INTEGRATION_URL"),
 	}
 }