+++ src/main/module/http/server/confg-server.go
@@ -4,14 +4,18 @@ import (
 	"go-api/src/main/container"
 	controllers "go-api/src/presentation/http/controllers"
 	v1_user_create "go-api/src/presentation/http/controllers/v1/users/create"
+	v1_user_delete "go-api/src/presentation/http/controllers/v1/users/delete"
 	v1_user "go-api/src/presentation/http/controllers/v1/users/find-by-di"
+	v1_user_grpc "go-api/src/presentation/http/controllers/v1/users/find-by-id-grpc"
 	"go-api/src/presentation/http/middlewares"
 )
 
 func loadControllers(container *container.Container) []controllers.Controller {
 	return []controllers.Controller{
 		v1_user.NewController(container),
+		v1_user_grpc.NewController((container)),
 		v1_user_create.NewController(container),
+		v1_user_delete.NewController(container),
 	}
 }