+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Ongoing Changes
 
+## [2.2.1]
+
+- fix refactoring integrations folders
+- add client to grpc 
+- add service to host grpc
+- update version go 1.18
+- fix changelog and add loggers
+- add new rote delete user
+- fix connection redis
+
 ## [2.0.0]
 - refectory interfaces in core
 - add work for manager cronjobs